+++ docroot/WEB-INF/service/de/ki/sbam/service/DifficultyLocalServiceUtil.java
@@ -41,7 +41,7 @@ public class DifficultyLocalServiceUtil {
 	 *
 	 * Never modify this class directly. Add custom service methods to {@link de.ki.sbam.service.impl.DifficultyLocalServiceImpl} and rerun ServiceBuilder to regenerate this class.
 	 */
-	public static de.ki.sbam.model.Difficulty addDifficultry(int score,
+	public static de.ki.sbam.model.Difficulty addDifficultry(long score,
 		boolean guaranteed) {
 		return getService().addDifficultry(score, guaranteed);
 	}
@@ -63,7 +63,7 @@ public class DifficultyLocalServiceUtil {
 	* @param score the primary key for the new difficulty
 	* @return the new difficulty
 	*/
-	public static de.ki.sbam.model.Difficulty createDifficulty(int score) {
+	public static de.ki.sbam.model.Difficulty createDifficulty(long score) {
 		return getService().createDifficulty(score);
 	}
 
@@ -85,7 +85,7 @@ public class DifficultyLocalServiceUtil {
 	* @return the difficulty that was removed
 	* @throws PortalException if a difficulty with the primary key could not be found
 	*/
-	public static de.ki.sbam.model.Difficulty deleteDifficulty(int score)
+	public static de.ki.sbam.model.Difficulty deleteDifficulty(long score)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService().deleteDifficulty(score);
 	}
@@ -177,7 +177,7 @@ public class DifficultyLocalServiceUtil {
 		return getService().dynamicQueryCount(dynamicQuery, projection);
 	}
 
-	public static de.ki.sbam.model.Difficulty fetchDifficulty(int score) {
+	public static de.ki.sbam.model.Difficulty fetchDifficulty(long score) {
 		return getService().fetchDifficulty(score);
 	}
 
@@ -221,7 +221,7 @@ public class DifficultyLocalServiceUtil {
 	* @return the difficulty
 	* @throws PortalException if a difficulty with the primary key could not be found
 	*/
-	public static de.ki.sbam.model.Difficulty getDifficulty(int score)
+	public static de.ki.sbam.model.Difficulty getDifficulty(long score)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService().getDifficulty(score);
 	}
